# mainServer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation of Alexandria's API. Endpoints can be tried out directly in this interactive documentation.",
        "title": "API documentation",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/articles/:articleID/requests/:requestID/threads": {
            "get": {
                "description": "Gets a list with all threads belonging to a specific request of an article",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all threads for a request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Thread"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperror.HTTPError"
                        }
                    }
                }
            }
        },
        "/articles/:articleID/thread/:threadType/id/:specificID/": {
            "post": {
                "description": "Creates thread entity, and specific thread entity. Returns id's of thread, specific thread and comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates thread entity",
                "parameters": [
                    {
                        "description": "Thread",
                        "name": "thread",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Thread"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ReturnThreadIds"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperror.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httperror.HTTPError"
                        }
                    }
                }
            }
        },
        "/articles/:articleID/versions/:versionID/history/:commitID/threads": {
            "get": {
                "description": "Gets a list with all threads belonging to a specific commit of an article",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all threads for a commit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Commit ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Thread"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperror.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/comments/thread/:threadID": {
            "post": {
                "description": "Save all types (commit/request/review) of comments to the database",
                "consumes": [
                    "application/json"
                ],
                "summary": "Save comment",
                "parameters": [
                    {
                        "description": "Comment",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Comment"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Thread ID",
                        "name": "threadID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "failed saving comment"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.Comment": {
            "type": "object",
            "required": [
                "authorId",
                "content",
                "creationDate"
            ],
            "properties": {
                "authorId": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "creationDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "threadId": {
                    "type": "integer"
                }
            }
        },
...
        "httperror.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
...
        "models.ReturnThreadIds": {
            "type": "object",
            "required": [
                "CommentId",
                "id",
                "threadId"
            ],
            "properties": {
                "CommentId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "threadId": {
                    "type": "integer"
                }
            }
        },
        "models.Thread": {
            "type": "object",
            "required": [
                "articleId",
                "comment"
            ],
            "properties": {
                "articleId": {
                    "type": "integer"
                },
                "comment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Comment"
                    }
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
    }
}